// openmetadata-ui/src/main/resources/ui/src/components/Settings/Services/ServiceConfig/FiltersConfigForm.types.ts
/*
 *  Copyright 2025 Collate.
 *  Licensed under the Apache License, Version 2.0 (the "License");
 *  you may not use this file except in compliance with the License.
 *  You may obtain a copy of the License at
 *  http://www.apache.org/licenses/LICENSE-2.0
 *  Unless required by applicable law or agreed to in writing, software
 *  distributed under the License is distributed on an "AS IS" BASIS,
 *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 *  See the License for the specific language governing permissions and
 *  limitations under the License.
 */
import { ComponentType, SVGProps } from 'react';

export type IconComponent = ComponentType<
  SVGProps<SVGSVGElement> & {
    size?: number;
  }
>;

export type FilterOperator =
  | 'contains'
  | 'endsWith'
  | 'is'
  | 'regex'
  | 'startsWith';

export interface FilterCondition {
  op: FilterOperator;
  sourceRegex?: string;
  value: string;
}

export interface FilterPatternConfig {
  excludes?: string[];
  includes?: string[];
}

export interface FilterSchemaProperty {
  default?: FilterPatternConfig;
  description?: string;
  javaType?: string;
  properties?: Record<string, unknown>;
  title?: string;
  type?: string;
}

export interface FilterSection {
  description?: string;
  fieldName: string;
  icon: IconComponent;
  label: string;
  singleLabel: string;
  systemExcludes: FilterCondition[];
}

export interface FilterSectionState {
  excludes: FilterCondition[];
  includes: FilterCondition[];
  restrict: boolean;
}

export type FiltersState = Record<string, FilterSectionState>;